refactor: add wechat info to accounts.ts file
